BE-7-add_validation

+++ app/routers/user.py
@@ -5,6 +5,9 @@ from sqlalchemy.ext.asyncio import AsyncSession
 from app.schemas.user_schemas import UserDetailResponse, UserUpdateRequest, UserResponse
 from app.services.user import UserService
 from app.db.database import get_session
+from app.db.user_model import User
+from app.routers.auth import get_current_user
+from app.services.exceptions import AccessDeniedError
 
 
 router = APIRouter(prefix="/users")
@@ -43,9 +46,15 @@ async def read_user_by_id(
 async def update_user(
     user_id: UUID,
     user_update: UserUpdateRequest,
+    current_user: User = Depends(get_current_user),
     session: AsyncSession = Depends(get_session),
     user_service: UserService = Depends(get_user_service),
 ):
+    if user_id != current_user.id:
+        raise AccessDeniedError(
+            "You are not allowed to update other users' information"
+        )
+
     user = await user_service.update_user(
         user_id=user_id, user_update=user_update, session=session
     )
@@ -56,9 +65,13 @@ async def update_user(
 @router.delete("/{user_id}", status_code=status.HTTP_204_NO_CONTENT)
 async def delete_user(
     user_id: UUID,
+    current_user: User = Depends(get_current_user),
     session: AsyncSession = Depends(get_session),
     user_service: UserService = Depends(get_user_service),
 ):
+    if user_id != current_user.id:
+        raise AccessDeniedError("You are not allowed to delete other users")
+
     await user_service.delete_user(user_id=user_id, session=session)
 
     return None
